Centos: added network call response checking, boost version check and numerous other clean up changes

+++ scripts/eosio_build_centos.sh
@@ -29,13 +29,13 @@
 
 	if [ $OS_VER -lt 7 ]; then
 		echo "You must be running Centos 7 or higher to install EOSIO."
-		echo "exiting now"
+		echo "exiting now."
 		exit 1
 	fi
 
 	if [ ${DISK_AVAIL%.*} -lt $DISK_MIN ]; then
 		echo "You must have at least ${DISK_MIN}GB of available storage to install EOSIO."
-		echo "exiting now"
+		echo "exiting now."
 		exit 1
 	fi
 	printf "\n\tChecking Yum installation\n"
@@ -176,10 +176,15 @@
 		printf "\tInstalling CMAKE\n"
 		mkdir -p ${HOME}/opt/ 2>/dev/null
 		cd ${HOME}/opt
-		curl -L -O https://cmake.org/files/v3.10/cmake-3.10.2.tar.gz
+		STATUS=$(curl -LO -w '%{http_code}' --connect-timeout 30 https://cmake.org/files/v3.10/cmake-3.10.2.tar.gz)
+		if [ "${STATUS}" -ne 200 ]; then
+			printf "\tUnable to download Boost libraries at this time.\n"
+			printf "\tExiting now.\n\n"
+			exit;
+		fi
 		tar xf cmake-3.10.2.tar.gz
 		rm -f cmake-3.10.2.tar.gz
-		ln -s cmake-3.10.2/ cmake
+		ln -s ${HOME}/opt/cmake-3.10.2/ cmake
 		cd cmake
 		./bootstrap
 		if [ $? -ne 0 ]; then
@@ -187,7 +192,7 @@
 			printf "\tExiting now.\n\n"
 			exit;
 		fi
-		make
+		make -j${CPU_CORE}
 		if [ $? -ne 0 ]; then
 			printf "\tError compiling CMAKE.\n"
 			printf "\tExiting now.\n\n"
@@ -198,36 +203,60 @@
 	fi
 
 	printf "\n\tChecking boost library installation.\n"
-	if [ ! -d ${HOME}/opt/boost_1_66_0 ]; then
-		printf "\tInstalling boost libraries\n"
+	BVERSION=`cat "${BOOST_ROOT}/include/boost/version.hpp" 2>/dev/null | grep BOOST_LIB_VERSION \
+	| tail -1 | tr -s ' ' | cut -d\  -f3 | sed 's/[^0-9\._]//gI'`
+	if [ "${BVERSION}" != "1_66" ]; then
+		printf "\tRemoving existing boost libraries in ${HOME}/opt/boost* .\n"
+		rm -rf ${HOME}/opt/boost*
+		if [ $? -ne 0 ]; then
+			printf "\n\tUnable to remove deprecated boost libraries at this time.\n"
+			printf "\n\tExiting now.\n"
+			exit 1
+		fi
+		printf "\tInstalling boost libraries.\n"
 		cd ${TEMP_DIR}
-		curl -L https://dl.bintray.com/boostorg/release/1.66.0/source/boost_1_66_0.tar.bz2 > boost_1.66.0.tar.bz2
-		tar xf boost_1.66.0.tar.bz2
-		cd boost_1_66_0/
+		STATUS=$(curl -LO -w '%{http_code}' --connect-timeout 30 https://dl.bintray.com/boostorg/release/1.66.0/source/boost_1_66_0.tar.bz2)
+		if [ "${STATUS}" -ne 200 ]; then
+			printf "\tUnable to download Boost libraries at this time.\n"
+			printf "\tExiting now.\n\n"
+			exit;
+		fi
+		tar xf ${TEMP_DIR}/boost_1_66_0.tar.bz2
+		rm -f  ${TEMP_DIR}/boost_1_66_0.tar.bz2
+		cd ${TEMP_DIR}/boost_1_66_0/
 		./bootstrap.sh "--prefix=$BOOST_ROOT"
+		if [ $? -ne 0 ]; then
+			printf "\n\tInstallation of boost libraries failed. 0\n"
+			printf "\n\tExiting now.\n"
+			exit 1
+		fi
 		./b2 install
+		if [ $? -ne 0 ]; then
+			printf "\n\tInstallation of boost libraries failed. 1\n"
+			printf "\n\tExiting now.\n"
+			exit 1
+		fi
 		rm -rf ${TEMP_DIR}/boost_1_66_0/
-		rm -f  ${TEMP_DIR}/boost_1.66.0.tar.bz2
 	else
-		printf "\tBoost 1.66 found at ${HOME}/opt/boost_1_66_0\n"
+		printf "\tBoost 1.66.0 found at ${HOME}/opt/boost_1_66_0.\n"
 	fi
 
 	printf "\n\tChecking MongoDB installation.\n"
     if [ ! -e ${MONGOD_CONF} ]; then
 		printf "\n\tInstalling MongoDB 3.6.3.\n"
 		cd ${HOME}/opt
-		curl -OL https://fastdl.mongodb.org/linux/mongodb-linux-x86_64-amazon-3.6.3.tgz
-		if [ $? -ne 0 ]; then
+		STATUS=$(curl -LO -w '%{http_code}' --connect-timeout 30 https://fastdl.mongodb.org/linux/mongodb-linux-x86_64-3.6.3.tgz)
+		if [ "${STATUS}" -ne 200 ]; then
 			printf "\tUnable to download MongoDB at this time.\n"
 			printf "\tExiting now.\n\n"
 			exit;
 		fi
-		tar xf mongodb-linux-x86_64-amazon-3.6.3.tgz
-		rm -f mongodb-linux-x86_64-amazon-3.6.3.tgz
-		ln -s ${HOME}/opt/mongodb-linux-x86_64-amazon-3.6.3/ ${HOME}/opt/mongodb
+		tar xf mongodb-linux-x86_64-3.6.3.tgz
+		rm -f mongodb-linux-x86_64-3.6.3.tgz
+		ln -s ${HOME}/opt/mongodb-linux-x86_64-3.6.3/ mongodb
 		mkdir ${HOME}/opt/mongodb/data
 		mkdir ${HOME}/opt/mongodb/log
-		touch ${HOME}/opt/mongodb/log/mongod.log
+		touch ${HOME}/opt/mongodb/log/mongodb.log
 		
 tee > /dev/null ${MONGOD_CONF} <<mongodconf
 systemLog:
@@ -250,8 +279,8 @@ mongodconf
     if [ ! -e /usr/local/lib/libmongocxx-static.a ]; then
 		printf "\n\tInstalling MongoDB C & C++ drivers.\n"
 		cd ${TEMP_DIR}
-		curl -LO https://github.com/mongodb/mongo-c-driver/releases/download/1.9.3/mongo-c-driver-1.9.3.tar.gz
-		if [ $? -ne 0 ]; then
+		STATUS=$(curl -LO -w '%{http_code}' --connect-timeout 30 https://github.com/mongodb/mongo-c-driver/releases/download/1.9.3/mongo-c-driver-1.9.3.tar.gz)
+		if [ "${STATUS}" -ne 200 ]; then
 			rm -f ${TEMP_DIR}/mongo-c-driver-1.9.3.tar.gz 2>/dev/null
 			printf "\tUnable to download MongoDB C driver at this time.\n"
 			printf "\tExiting now.\n\n"
@@ -318,6 +347,11 @@ mongodconf
 		printf "\tInstalling secp256k1-zkp (Cryptonomex branch)\n"
 		cd ${TEMP_DIR}
 		git clone https://github.com/cryptonomex/secp256k1-zkp.git
+		if [ $? -ne 0 ]; then
+			printf "\tUnable to clone repo secp256k1-zkp @ https://github.com/cryptonomex/secp256k1-zkp.git.\n"
+			printf "\tExiting now.\n\n"
+			exit;
+		fi
 		cd secp256k1-zkp
 		./autogen.sh
 		if [ $? -ne 0 ]; then
@@ -338,26 +372,6 @@ mongodconf
 		printf "\tsecp256k1 found\n"
 	fi
 	
-	printf "\n\tChecking binaryen installation.\n"
-	if [ ! -d ${HOME}/opt/binaryen ]; then
-		printf "\tInstalling binaryen v1.37.14:\n"
-		cd ${TEMP_DIR}
-		git clone https://github.com/EOSIO/binaryen
-		cd binaryen
-		git checkout eosio
-		$CMAKE . && make
-		if [ $? -ne 0 ]; then
-			printf "\tError compiling binaryen.\n"
-			printf "\tExiting now.\n\n"
-			exit;
-		fi
-		mkdir -p ${HOME}/opt/binaryen/ 2>/dev/null
-		mv ${TEMP_DIR}/binaryen/bin ${HOME}/opt/binaryen/
-		rm -rf ${TEMP_DIR}/binaryen
-	else
-		printf "\tBinaryen found at ${HOME}/opt/binaryen\n"
-	fi
-
 	printf "\n\tChecking LLVM with WASM support installation.\n"
 	if [ ! -d ${HOME}/opt/wasm/bin ]; then
 		printf "\tInstalling LLVM & WASM\n"
@@ -365,8 +379,18 @@ mongodconf
 		mkdir llvm-compiler  2>/dev/null
 		cd llvm-compiler
 		git clone --depth 1 --single-branch --branch release_40 https://github.com/llvm-mirror/llvm.git
+		if [ $? -ne 0 ]; then
+			printf "\tUnable to clone llvm repo @ https://github.com/llvm-mirror/llvm.git.\n"
+			printf "\tExiting now.\n\n"
+			exit;
+		fi
 		cd llvm/tools
 		git clone --depth 1 --single-branch --branch release_40 https://github.com/llvm-mirror/clang.git
+		if [ $? -ne 0 ]; then
+			printf "\tUnable to clone clang repo @ https://github.com/llvm-mirror/clang.git.\n"
+			printf "\tExiting now.\n\n"
+			exit;
+		fi
 		cd ..
 		mkdir build 2>/dev/null
 		cd build
@@ -378,7 +402,7 @@ mongodconf
 			printf "\tExiting now.\n\n"
 			exit;
 		fi
-		make -j$(nproc)
+		make -j${CPU_CORE}
 		if [ $? -ne 0 ]; then
 			printf "\tError compiling LLVM and clang with EXPERIMENTAL WASM support.\n"
 			printf "\tExiting now.\n\n"
